second cut: wip
