+++ cmd/apptable/Cargo.toml
@@ -1,0 +1,11 @@
+[package]
+name = "apptable"
+version = "0.1.0"
+edition = "2021"
+
+[dependencies]
+humility = { path = "../../humility-core", package = "humility-core" }
+humility-cmd = { path = "../../humility-cmd" }
+structopt = "0.3"
+anyhow = { version = "1.0.44", features = ["backtrace"] }
+
