+++ humility-core/src/hubris.rs
@@ -2,7 +2,7 @@
 // License, v. 2.0. If a copy of the MPL was not distributed with this
 // file, You can obtain one at https://mozilla.org/MPL/2.0/.
 
-use crate::debug::*;
+use crate::arch::ARMRegister;
 use capstone::prelude::*;
 use indexmap::IndexMap;
 use serde::Deserialize;
