+++ humility-core/src/lib.rs
@@ -1,0 +1,13 @@
+// This Source Code Form is subject to the terms of the Mozilla Public
+// License, v. 2.0. If a copy of the MPL was not distributed with this
+// file, You can obtain one at https://mozilla.org/MPL/2.0/.
+
+pub mod hubris;
+pub mod core;
+pub mod arch;
+
+#[macro_use]
+extern crate num_derive;
+
+#[macro_use]
+extern crate log;
