+++ src/cmd.rs
@@ -2,7 +2,8 @@
 // License, v. 2.0. If a copy of the MPL was not distributed with this
 // file, You can obtain one at https://mozilla.org/MPL/2.0/.
 
-mod apptable;
+// mod apptable;
+/*
 mod diagnose;
 mod dump;
 mod etm;
@@ -28,59 +29,17 @@ mod stmsecure;
 mod tasks;
 mod test;
 mod trace;
+*/
 
-use crate::Args;
+use humility_cmd::Args;
 use crate::{attach_dump, attach_live};
 use anyhow::{bail, Result};
 use humility::core::Core;
 use humility::hubris::*;
+use humility_cmd::{Archive, Attach, Validate, Command};
 use std::collections::HashMap;
 use std::convert::TryInto;
 use structopt::clap::App;
-
-#[allow(dead_code)]
-#[derive(Copy, Clone, Debug)]
-pub enum Archive {
-    Required,
-    Optional,
-    Prohibited,
-}
-
-#[allow(dead_code)]
-#[derive(Copy, Clone, Debug)]
-pub enum Attach {
-    LiveOnly,
-    DumpOnly,
-    Any,
-}
-
-#[allow(dead_code)]
-#[derive(Copy, Clone, Debug)]
-pub enum Validate {
-    Match,
-    Booted,
-    None,
-}
-
-pub enum Command {
-    Attached {
-        name: &'static str,
-        archive: Archive,
-        attach: Attach,
-        validate: Validate,
-        run: fn(
-            &mut HubrisArchive,
-            &mut dyn Core,
-            &Args,
-            &Vec<String>,
-        ) -> Result<()>,
-    },
-    Unattached {
-        name: &'static str,
-        archive: Archive,
-        run: fn(&mut HubrisArchive, &Args, &Vec<String>) -> Result<()>,
-    },
-}
 
 pub fn init<'a, 'b>(
     app: App<'a, 'b>,
@@ -89,7 +48,8 @@ pub fn init<'a, 'b>(
     let mut rval = app;
 
     let dcmds = [
-        apptable::init,
+        cmd_apptable::init,
+        /*
         diagnose::init,
         dump::init,
         etm::init,
@@ -115,6 +75,7 @@ pub fn init<'a, 'b>(
         test::init,
         trace::init,
         stmsecure::init,
+        */
     ];
 
     for dcmd in &dcmds {
@@ -188,86 +149,5 @@ pub fn subcommand(
         }
     } else {
         bail!("command {} not found", subargs[0]);
-    }
-}
-
-fn printmem(bytes: &[u8], addr: u32, size: usize, width: usize) {
-    let mut addr = addr;
-
-    let print = |line: &[u8], addr, offs| {
-        print!("0x{:08x} | ", addr);
-
-        for i in (0..width).step_by(size) {
-            if i < offs || i - offs >= line.len() {
-                print!(" {:width$}", "", width = size * 2);
-                continue;
-            }
-
-            let slice = &line[i - offs..i - offs + size];
-
-            print!(
-                "{:0width$x} ",
-                match size {
-                    1 => line[i - offs] as u32,
-                    2 => u16::from_le_bytes(slice.try_into().unwrap()) as u32,
-                    4 => u32::from_le_bytes(slice.try_into().unwrap()) as u32,
-                    _ => {
-                        panic!("invalid size");
-                    }
-                },
-                width = size * 2
-            );
-        }
-
-        print!("| ");
-
-        for i in 0..width {
-            if i < offs || i - offs >= line.len() {
-                print!(" ");
-            } else {
-                let c = line[i - offs] as char;
-
-                if c.is_ascii() && !c.is_ascii_control() {
-                    print!("{}", c);
-                } else {
-                    print!(".");
-                }
-            }
-        }
-
-        println!();
-    };
-
-    let offs = (addr & (width - 1) as u32) as usize;
-    addr -= offs as u32;
-
-    /*
-     * Print out header line, OpenBoot PROM style
-     */
-    print!("  {:8}  ", "");
-
-    for i in (0..width).step_by(size) {
-        if i == offs {
-            print!(" {:>width$}", "\\/", width = size * 2);
-        } else {
-            print!(" {:>width$x}", i, width = size * 2);
-        }
-    }
-
-    println!();
-
-    /*
-     * Print our first line.
-     */
-    let lim = std::cmp::min(width - offs, bytes.len());
-    print(&bytes[0..lim], addr, offs);
-
-    if lim < bytes.len() {
-        let lines = bytes[lim..].chunks(width);
-
-        for line in lines {
-            addr += width as u32;
-            print(line, addr, 0);
-        }
     }
 }
