+++ src/cmd/itm.rs
@@ -4,14 +4,14 @@
 
 use crate::attach_live;
 use crate::cmd::{Archive, Command};
-use crate::dwt::*;
-use crate::itm::*;
-use crate::scs::*;
-use crate::tpiu::*;
+use humility_cortex::dwt::*;
+use humility_cortex::itm::*;
+use humility_cortex::scs::*;
+use humility_cortex::tpiu::*;
 use crate::Args;
 use anyhow::{bail, Result};
 use humility::core::Core;
-use humility::debug::*;
+use humility_cortex::debug::*;
 use humility::hubris::*;
 use std::fs::File;
 use std::io::Read;
