+++ src/cmd/trace.rs
@@ -3,7 +3,7 @@
 // file, You can obtain one at https://mozilla.org/MPL/2.0/.
 
 use crate::cmd::*;
-use crate::itm::*;
+use humility_cortex::itm::*;
 use crate::Args;
 use anyhow::{anyhow, Result};
 use humility::core::Core;
